pyproject.toml
[project]
name = "lang-chain-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Ankur Dutta", email = "ankur.611@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain-openai (>=0.3.4,<0.4.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "black (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "langchain-pinecone (>=0.2.3,<0.3.0)",
    "isort (>=6.0.0,<7.0.0)",
    "langgraph (>=0.2.73,<0.3.0)"
]

[tool.poetry]
packages = [
    {include = "app"},
    {include = "rag"},
    {include = "reflection-agent"},
    {include = "reflexion-agent"}
]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
langchain = "^0.3.18"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
